+++ pkg/config/network.go
@@ -1,4 +1,4 @@
-package app
+package config
 
 import (
 	"bytes"
