+++ pkg/config/networks/coreum-devnet-1/validator-0.json
