+++ pkg/config/networks/coreum-devnet-1/validator-1.json
