+++ pkg/config/networks/coreum-devnet-1/validator-2.json
