+++ pkg/config/networks/coreum-devnet-1/validator-3.json
