+++ pkg/config/prefix.go
@@ -1,4 +1,4 @@
-package app
+package config
 
 import sdk "github.com/cosmos/cosmos-sdk/types"
